jaded-rose-chatbot/requirements.txt
# Jaded Rose Chatbot — Python dependencies

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
websockets>=12.0

# Telegram
python-telegram-bot>=20.7

# WhatsApp (Twilio)
twilio>=9.0.0

# Gmail API
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.114.0

# AI / Embeddings
openai>=1.12.0
pinecone-client>=3.0.0

# Redis (conversation memory)
redis>=5.0.0

# HTTP client (carrier APIs)
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.9
